+++ lib/src/util/analytics/analytics_util.dart
@@ -1,0 +1,15 @@
+abstract class ImpaktfullAnalyticsUtil {
+  Future<void> init({
+    bool enabled = false,
+  });
+
+  Future<void> setEnabled(bool enabled);
+
+  void logEvent(String eventName, {Map<String, dynamic>? data});
+
+  void setUserProperty(String property, String value);
+
+  void setUserId(String userId);
+
+  void setScreenName(String screenName);
+}