+++ lib/src/util/crash_reporting/crash_reporting_util.dart
@@ -1,0 +1,17 @@
+abstract class ImpaktfullCrashReportingUtil {
+  Future<void> init({bool enabled = false});
+
+  Future<void> setEnabled(bool enabled);
+
+  void setUserId(String userId);
+
+  void setUserProperty(String property, String value);
+
+  void logError({
+    required Object error,
+    required StackTrace stackTrace,
+    String? message,
+  });
+
+  void log(String message, {Map<String, dynamic>? data});
+}